added pagination to /games api endpoint, 30 results per page

+++ back_end/Project/src/main/java/com/GamBox/Project/dto/RequestPageSize.java
@@ -1,4 +1,4 @@
 package com.GamBox.Project.dto;
 
-public record RequestPageSize(int page, int size) {
+public record RequestPageSize(Long page, int size) {
 }
